feat(checksum): extract utility functions to specific file to make logic simpler
