+++ src/uucore/src/lib/features/checksum/mod.rs
@@ -13,8 +13,8 @@ use std::{
     ffi::{OsStr, OsString},
     fs::File,
     io::{self, BufReader, Read},
-    path::Path,
 };
+use utils::{get_filename_for_output, unescape_filename};
 
 use crate::{
     error::{set_exit_code, FromIo, UError, UResult, USimpleError},
@@ -27,6 +27,8 @@ use std::io::stdin;
 use thiserror::Error;
 
 pub mod algo;
+mod utils;
+pub use utils::escape_filename; // Used in hashsum.rs
 
 pub const ALGORITHM_OPTIONS_SYSV: &str = "sysv";
 pub const ALGORITHM_OPTIONS_BSD: &str = "bsd";
@@ -87,6 +89,33 @@ struct ChecksumResult {
     pub failed_open_file: i32,
 }
 
+impl ChecksumResult {
+    /// Print diagnostic lines at the end of the processing of a checksum file.
+    #[allow(clippy::comparison_chain)]
+    fn print_output(&self, ignore_missing: bool, status: bool) {
+        if self.bad_format == 1 {
+            show_warning_caps!("{} line is improperly formatted", self.bad_format);
+        } else if self.bad_format > 1 {
+            show_warning_caps!("{} lines are improperly formatted", self.bad_format);
+        }
+
+        if !status {
+            if self.failed_cksum == 1 {
+                show_warning_caps!("{} computed checksum did NOT match", self.failed_cksum);
+            } else if self.failed_cksum > 1 {
+                show_warning_caps!("{} computed checksums did NOT match", self.failed_cksum);
+            }
+        }
+        if !ignore_missing {
+            if self.failed_open_file == 1 {
+                show_warning_caps!("{} listed file could not be read", self.failed_open_file);
+            } else if self.failed_open_file > 1 {
+                show_warning_caps!("{} listed files could not be read", self.failed_open_file);
+            }
+        }
+    }
+}
+
 #[derive(Debug, Error)]
 pub enum ChecksumError {
     #[error("the --raw option is not supported with multiple files")]
@@ -129,30 +158,6 @@ impl UError for ChecksumError {
     }
 }
 
-#[allow(clippy::comparison_chain)]
-fn cksum_output(res: &ChecksumResult, ignore_missing: bool, status: bool) {
-    if res.bad_format == 1 {
-        show_warning_caps!("{} line is improperly formatted", res.bad_format);
-    } else if res.bad_format > 1 {
-        show_warning_caps!("{} lines are improperly formatted", res.bad_format);
-    }
-
-    if !status {
-        if res.failed_cksum == 1 {
-            show_warning_caps!("{} computed checksum did NOT match", res.failed_cksum);
-        } else if res.failed_cksum > 1 {
-            show_warning_caps!("{} computed checksums did NOT match", res.failed_cksum);
-        }
-    }
-    if !ignore_missing {
-        if res.failed_open_file == 1 {
-            show_warning_caps!("{} listed file could not be read", res.failed_open_file);
-        } else if res.failed_open_file > 1 {
-            show_warning_caps!("{} listed files could not be read", res.failed_open_file);
-        }
-    }
-}
-
 // Regexp to handle the three input formats:
 // 1. <algo>[-<bits>] (<filename>) = <checksum>
 //    algo must be uppercase or b (for blake2b)
@@ -165,16 +170,6 @@ const DOUBLE_SPACE_REGEX: &str = r"^(?P<checksum>[a-fA-F0-9]+)\s{2}(?P<filename>
 
 // In this case, we ignore the *
 const SINGLE_SPACE_REGEX: &str = r"^(?P<checksum>[a-fA-F0-9]+)\s(?P<filename>\*?.*)$";
-
-fn get_filename_for_output(filename: &OsStr, input_is_stdin: bool) -> String {
-    if input_is_stdin {
-        "standard input"
-    } else {
-        filename.to_str().unwrap()
-    }
-    .maybe_quote()
-    .to_string()
-}
 
 /// Determines the appropriate regular expression to use based on the provided lines.
 fn determine_regex(lines: &[OsString]) -> Option<(Regex, bool)> {
@@ -559,7 +554,7 @@ where
         }
 
         // if any incorrectly formatted line, show it
-        cksum_output(&res, opts.ignore_missing, opts.status);
+        res.print_output(opts.ignore_missing, opts.status);
     }
 
     Ok(())
@@ -628,84 +623,11 @@ pub fn calculate_blake2b_length(length: usize) -> UResult<Option<usize>> {
     }
 }
 
-pub fn unescape_filename(filename: &[u8]) -> (Vec<u8>, &'static str) {
-    let mut unescaped = Vec::with_capacity(filename.len());
-    let mut byte_iter = filename.iter().peekable();
-    loop {
-        let Some(byte) = byte_iter.next() else {
-            break;
-        };
-        if *byte == b'\\' {
-            match byte_iter.next() {
-                Some(b'\\') => unescaped.push(b'\\'),
-                Some(b'n') => unescaped.push(b'\n'),
-                Some(b'r') => unescaped.push(b'\r'),
-                Some(x) => {
-                    unescaped.push(b'\\');
-                    unescaped.push(*x);
-                }
-                _ => {}
-            }
-        } else {
-            unescaped.push(*byte);
-        }
-    }
-    let prefix = if unescaped == filename { "" } else { "\\" };
-    (unescaped, prefix)
-}
-
-pub fn escape_filename(filename: &Path) -> (String, &'static str) {
-    let original = filename.as_os_str().to_string_lossy();
-    let escaped = original
-        .replace('\\', "\\\\")
-        .replace('\n', "\\n")
-        .replace('\r', "\\r");
-    let prefix = if escaped == original { "" } else { "\\" };
-    (escaped, prefix)
-}
-
 #[cfg(test)]
 mod tests {
     use std::ffi::OsString;
 
     use super::*;
-
-    #[test]
-    fn test_unescape_filename() {
-        let (unescaped, prefix) = unescape_filename(b"test\\nfile.txt");
-        assert_eq!(unescaped, b"test\nfile.txt");
-        assert_eq!(prefix, "\\");
-        let (unescaped, prefix) = unescape_filename(b"test\\nfile.txt");
-        assert_eq!(unescaped, b"test\nfile.txt");
-        assert_eq!(prefix, "\\");
-
-        let (unescaped, prefix) = unescape_filename(b"test\\rfile.txt");
-        assert_eq!(unescaped, b"test\rfile.txt");
-        assert_eq!(prefix, "\\");
-
-        let (unescaped, prefix) = unescape_filename(b"test\\\\file.txt");
-        assert_eq!(unescaped, b"test\\file.txt");
-        assert_eq!(prefix, "\\");
-    }
-
-    #[test]
-    fn test_escape_filename() {
-        let (escaped, prefix) = escape_filename(Path::new("testfile.txt"));
-        assert_eq!(escaped, "testfile.txt");
-        assert_eq!(prefix, "");
-
-        let (escaped, prefix) = escape_filename(Path::new("test\nfile.txt"));
-        assert_eq!(escaped, "test\\nfile.txt");
-        assert_eq!(prefix, "\\");
-
-        let (escaped, prefix) = escape_filename(Path::new("test\rfile.txt"));
-        assert_eq!(escaped, "test\\rfile.txt");
-        assert_eq!(prefix, "\\");
-
-        let (escaped, prefix) = escape_filename(Path::new("test\\file.txt"));
-        assert_eq!(escaped, "test\\\\file.txt");
-        assert_eq!(prefix, "\\");
-    }
 
     #[test]
     fn test_calculate_blake2b_length() {
